feat: setup notification service using nestjs

+++ solidity/contracts/PawningShop.sol
@@ -476,7 +476,7 @@ contract PawningShop {
     }
 
     function createBid(
-        uint8 rate,
+        uint256 rate,
         uint256 duration,
         bool isInterestProRated,
         uint256 pawnId
